Initial commit. Copied from portions of svn-ece3724 r2211.
This repo is a merge of the pic24lib repo, containing only the core library,
and the remainder of the files native to this library. These files gives
examples of library usage, provide build support, and contain library-related
tools.

The intention is to pull any updates to the core library files from pic24lib,
and to provide a source of updates from which pic24lib can pull. Therefore, all
commits should modify just core library files merged from pic24lib or just
the native files in this repo, not both, to enable this type of pull.